+++ src/main/java/com/brand/blockus/datagen/providers/BlockusLangEnProvider.java
@@ -94,6 +94,9 @@ public class BlockusLangEnProvider extends FabricLanguageProvider {
             addWool(translationBuilder, BlockusBlocks.PATTERNED_WOOL, color, name(color.getName()) + " Patterned Wool");
             addWool(translationBuilder, BlockusBlocks.GINGHAM_WOOL, color, name(color.getName()) + " Gingham Wool");
         }
+        translationBuilder.add(BlockusBlocks.RAINBOW_WOOL, "Rainbow Wool");
+        translationBuilder.add(BlockusBlocks.RAINBOW_CARPET, "Rainbow Carpet");
+        translationBuilder.add(BlockusBlocks.RAINBOW_BED, "Rainbow Bed");
 
         // Basalt
         addBssw(translationBuilder, BlockusBlocks.ROUGH_BASALT, "Rough Basalt");