+++ src/main/java/com/brand/blockus/registry/content/BlockusBlocks.java
@@ -530,6 +530,9 @@ public class BlockusBlocks extends BlockFactory {
     public static final Block TINTED_BEVELED_GLASS = registerCopy("tinted_beveled_glass", TintedGlassBlock::new, Blocks.TINTED_GLASS, properties -> properties.isValidSpawn(BlockusBlocks::never).isRedstoneConductor(BlockusBlocks::never).isSuffocating(BlockusBlocks::never).isViewBlocking(BlockusBlocks::never));
 
     // Patterned Wools
+    public static final Block RAINBOW_WOOL = dyedBlock("rainbow_wool", DyeColor.YELLOW, (var0, p) -> new Block(p), Blocks.WOOL.white());
+    public static final Block RAINBOW_CARPET = dyedBlock("rainbow_carpet", DyeColor.YELLOW, WoolCarpetBlock::new, Blocks.CARPET.white());
+    public static final Block RAINBOW_BED = dyedBlock("rainbow_bed", DyeColor.YELLOW, BedBlock::new, Blocks.BED.white());
     public static final WoolBundle PATTERNED_WOOL = WoolBundle.register(BlockusIds.PATTERNED_WOOL, BlockusIds.PATTERNED_WOOL_STAIRS, BlockusIds.PATTERNED_WOOL_SLAB, BlockusIds.PATTERNED_CARPET, BlockBuilder::registerBlock);
     public static final WoolBundle GINGHAM_WOOL = WoolBundle.register(BlockusIds.GINGHAM_WOOL, BlockusIds.GINGHAM_WOOL_STAIRS, BlockusIds.GINGHAM_WOOL_SLAB, BlockusIds.GINGHAM_CARPET, BlockBuilder::registerBlock);