+++ changelog.md
@@ -4,4 +4,5 @@
 * Added Polished Sulfur Pressure Plate & Button
 * Added Polished Cinnabar Pressure Plate & Button
 * Added/modified some tags for Sulfur Cube archetypes
+* Added Rainbow Wool, Carpet & Bed